[step2] Reward, LottoResult 별도 파일로 분리
